Fix remaining audit bugs: security, UX, performance, data integrity
- Remove forced login gate on /booking (Pricing) page; show login prompt
  inline instead so users can see the page before authenticating
- Fix article slug collision: article 4 had article 1's slug
- Add rate limiting (10 req/15 min) on Google OAuth exchange endpoint
- Eliminate N+1 queries in getCourse and getAdminCourse using JOIN queries
  and a reshapeCourseRows helper; reduces DB round-trips from O(N modules
  * M lessons) to a single query per endpoint
- Validate video URLs against YouTube/Vimeo allowlist in createLesson and
  updateLesson; prevent arbitrary iframe injection
- Add format validation (email, YYYY-MM-DD date, HH:MM time, name length,
  future-date check) in bookSession calendar controller
- Fix slot conflict check using 1-hour window to match calendar event
  duration (was 30 min, causing double-booking of second half of sessions)
- Cast COUNT() to int in enrollment query; was returning strings from pg
- Add CHECK (price >= 0) to schema; apply to live DB via ALTER TABLE

+++ server/controllers/enrollmentController.js
@@ -53,8 +53,8 @@ export async function getMyEnrollments(req, res) {
     const { rows } = await pool.query(
       `SELECT c.id, c.slug, c.title, c.description, c.thumbnail_url, c.price,
               e.enrolled_at,
-              COUNT(DISTINCT l.id) AS total_lessons,
-              COUNT(DISTINCT p.lesson_id) AS completed_lessons
+              COUNT(DISTINCT l.id)::int AS total_lessons,
+              COUNT(DISTINCT p.lesson_id)::int AS completed_lessons
        FROM enrollments e
        JOIN courses c ON c.id = e.course_id
        LEFT JOIN modules m ON m.course_id = c.id
